reckon
Enum Term

pub enum Term {
    Sym(Symbol),
    Var(Var),
    App(AppTerm),
    Int(i64),
    Str(String),
    Nil,
    True,
    False,
    Cons(Box<Term>, Box<Term>),
    Complement(Box<Term>),
    Cut,
}

Representation of a logical term.

A term is a tree structure, where each node is a term.

There are literal terms, like integers, strings, and booleans. These are leaf nodes in the term tree.

There are also compound terms, like application terms and cons cells. These are nodes with children in the term tree, and are used to represent more complex structures which can be used in logical rules.

Variables are used to represent yet unknown terms, and can be bound to other terms. Symbols are used to represent non-variable terms in logical expressions, but with a unique, string interned identifier.

Variants

Sym(Symbol)

A non-variable symbol “x”

Var(Var)

A variable term “?x”

App(AppTerm)

An application term

Int(i64)

A signed integer term

Str(String)

A string term

Nil

A nil term

True

Truth

False

Falsehood

Cons(Box<Term>, Box<Term>)

A cons cell

This is a list cell with a head and a tail.

Complement(Box<Term>)

A complement term

This is a term that is negated.

Cut

The special built-in cut predicate.

Evaluating it prunes all further choices for the currently active rule.

Implementations

impl Term

pub fn size(&self) -> usize

Get the size of this term.

The size of a term is the number of nodes in the term tree.

This is useful for creating a heuristic for selecting which terms to unify first.

pub fn var(name: impl ToString) -> Self

Create a new symbol term.

pub fn app(func: impl ToString, args: Vec<Term>) -> Self

Create a new application term.

pub fn negate(&self) -> Self

Complement this term.

If this term is a boolean term, it will be negated. If this term is already a complement, the complement will be removed. All other terms will be wrapped in a complement.

pub fn unify<S>( &self, other: &Self, env: &mut Env<S>, negated: bool, ) -> Result<Self, (Self, Self)>
where S: Solver,

Unify this term with another term.

This will attempt to bind the variable terms in this term to the corresponding terms in the other term.

If the structures of the terms are incompatible, an error is returned.

The self term will not be modified, and a new term will be returned. For in-place unification, use unify_in_place.

pub fn unify_in_place<S>( &mut self, other: &Self, env: &mut Env<S>, negated: bool, ) -> Result<(), (Self, Self)>
where S: Solver,

Unify this term with another term.

This will attempt to bind the variable terms in this term to the corresponding terms in the other term.

If the structures of the terms are incompatible, an error is returned, and the environment is not modified.

pub fn is_app(&self) -> bool

Is this term an application term?

pub fn has_vars(&self) -> bool

Does this term contain any variables?

pub fn used_vars(&self, vars: &mut HashSet<Var>)

Get the set of variables used in this term.

The variables used in this term are collected and inserted into the given set.

pub fn reduce<S>(&self, env: &Env<S>) -> Self
where S: Solver,

Return a new, reduced term.

This will apply the environment’s variable bindings to this term, performing variable substitution and simplification.

Complements of boolean terms will be simplified to their negation.

pub fn reduce_in_place<S>(&mut self, env: &Env<S>)
where S: Solver,

Reduce this term in place.

This will apply the environment’s variable bindings to this term, performing variable substitution and simplification.

Complements of boolean terms will be simplified to their negation.

pub fn subterms(&self) -> Vec<&Self>

Get references to the subterms of this term. This will only return references to the subterms of this term, not the term itself. Also, it only returns references to the immediate subterms, not the subterms of subterms.

pub fn subterms_mut(&mut self) -> Vec<&mut Self>

Get mutable references to the subterms of this term. This will only return mutable references to the subterms of this term, not the term itself. Also, it only returns references to the immediate subterms, not the subterms of subterms.

pub fn traverse(&self, f: &mut impl FnMut(&Self) -> bool)

Apply a function that inspects this term, and all subterms recursively.

When the function returns false, the traversal stops. When the function returns true, the traversal continues.

pub fn traverse_mut(&mut self, f: &mut impl FnMut(&mut Self) -> bool)

Apply a function that mutates this term, and all subterms recursively.

When the function returns false, the traversal stops. When the function returns true, the traversal continues.

pub fn substitute(&mut self, bindings: &HashMap<Var, Term>)

Substitute all occurrences of variables in this term with the given bindings.

This will modify this term in place, replacing all occurrences of variables with their corresponding terms in the bindings.

pub fn substitute_var(&mut self, var: Var, term: Term)

Substitute a single variable with a term.

This will modify this term in place, replacing all occurrences of the given variable with the given term.

pub fn has_application_of(&self, app_term: &AppTerm) -> bool

Does this term contain an application of the given application term?

If the term contains an application of a function with the same name as the given application term, this function returns true.

Otherwise, it returns false.