reckon
Struct Env

pub struct Env<S>
where
    S: Solver,
{ /* private fields */ }

An environment for a solver.

This stores the rules, variable bindings, search configuration, and solver for a given search.

The environment is used to store the state of the search, store variable bindings for unification, and to perform the search for solutions.

It also stores some bookkeeping information, such as the number of steps performed in the search.

Implementations

impl<S> Env<S>
where S: Solver,

pub fn new(rules: &[Rule]) -> Self

Create a new environment with the given rules.

pub fn reset_solver(&mut self)

Reset the solver in the environment.

This will clear the memoization cache in the solver.

pub fn reset(&mut self)

Clear the environment, removing all rules and variable bindings.

pub fn get_steps(&self) -> usize

Get the number of steps performed in the search.

pub fn reset_steps(&mut self)

Reset the number of steps performed in the search.

pub fn search_config_mut(&mut self) -> &mut SearchConfig<S>

Get the search configuration for the environment.

pub fn with_search_config(self, search_config: &SearchConfig<S>) -> Self

Supply a new search configuration for the environment.

pub fn used_vars(&self, vars: &mut HashSet<Var>)

Get the variables used in the environment.

pub fn prune_redundant_variables(&mut self) -> HashMap<Var, Term>

Remove useless variables.

If X is bound to Y, and Y is bound to Z, then X is bound to Z.

pub fn prune(&mut self, original_query: &Query, query: &mut Query)

Prune the environment, removing unused variables.

It takes the full, original query supplied by the user, and the current query that is being searched.

pub fn var_bindings(&self) -> &HashMap<Var, Term>

Get the variable bindings for the environment.

pub fn var_bindings_mut(&mut self) -> &mut HashMap<Var, Term>

Get the mutable variable bindings for the environment.

pub fn set_var(&mut self, var: Var, term: Term)

Set a variable in the environment.

pub fn get_var(&self, var: Var) -> Option<&Term>

Get a variable from the environment.

pub fn get_rules(&self) -> &[Rule]

Get the rules for the environment.

pub fn add_rule(&mut self, rule: Rule)

Add a rule to the environment.

pub fn to_full_solution( &self, original_query: &Query, final_query: &Query, ) -> Result<Solution, HashSet<Term>>

Convert the bindings in this environment to a solution.

This will take all the variables in the query, and return the reduced bindings for those variables.

If there are variables in the query that are not bound, this will return an error with the set of unbound variables.

pub fn to_partial_solution( &self, original_query: &Query, final_query: &Query, ) -> Solution

Convert the bindings in this environment to a partial solution.

This will take all the variables in the query, and return the reduced bindings for those variables.

If all the variables are bound, this is a full solution. If there are variables in the query that are not bound, this will return the partial solution with as many variables bound as possible.

pub fn apply_rules(&mut self, goal: &Term, query: &mut Query) -> bool

Apply all the rules in the environment to the goal term, adding the new goals to the query.

pub fn apply_rules_to_query(&mut self, query: &mut Query) -> bool

Apply all the rules in the environment to all the goals query, adding the new goals to the query.

If the search configuration requires a head match, the rules will be applied to the first goal in the query before applying them to the rest of the goals. If the application fails, the rest of the goals will not be considered.

If the search configuration does not require a head match, the rules will always be applied to all the goals in the query.

pub fn apply_rule( &mut self, rule: usize, goal: &Term, query: &mut Query, ) -> bool

Apply a rule to a goal term, adding the new goals to the query.

pub fn apply_rule_to_query(&mut self, rule: usize, query: &mut Query) -> bool

Apply a rule to a query, adding the new goals to the query.

If the search configuration requires a head match, the rule will be applied to the first goal in the query before applying it to the rest of the goals. If the application fails, the rest of the goals will not be considered.

If the search configuration does not require a head match, the rule will always be applied to all the goals in the query.

pub fn prove_true(&mut self, query: &Query) -> Result<Solution, HashSet<Term>>

Attempt to find a solution for the query using the search configuration.

If a solution is found, it will be returned. If no solution is found, an error will be returned with the set of unbound variables.

pub fn prove_false(&mut self, query: &Query) -> Result<(), HashSet<Term>>

Attempt to prove that the query is false using the search configuration.

If the query is false, it will be added to the set of proven false queries, and the function will return Ok(()).

If the query is true, an error will be returned with the set of unsolved goals.

pub fn find_solutions( &mut self, original_query: &Query, ) -> Result<HashSet<Solution>, HashSet<Term>>

Find solutions for the query using the given search configuration.

pub fn find_solutions_dfs( &mut self, original_query: &Query, max_solutions: usize, ) -> Result<HashSet<Solution>, HashSet<Term>>

Perform a DFS search for solutions, using the given search configuration.

pub fn find_solutions_bfs( &mut self, original_query: &Query, max_solutions: usize, ) -> Result<HashSet<Solution>, HashSet<Term>>

Perform a BFS search for solutions, using the given search configuration.