reckon
Struct Query

pub struct Query { /* private fields */ }

A query to provide to the solver to prove true or false.

A query is a list of goals that the solver must prove true. If the solver can prove all the goals true, then the query is true.

Implementations

impl Query

pub fn new(goals: impl IntoIterator<Item = Term>) -> Self

Create a new query with the given goals to prove

pub fn empty() -> Self

Create an empty query, which is interpreted as true

pub fn size(&self) -> usize

Get the sum of the sizes of all the goals in the query

The size of a goal term is the total number of subterms in the term

pub fn remove_irreducible_negatives_in_place(&mut self, env: &Env<impl Solver>)

Remove all the complemented goals that cannot be simplified by any of the rules in the environment.

These are inherently negative goals that cannot be proven

pub fn has_complements(&self) -> bool

Does the query contain any complements?

pub fn has_only_complements(&self) -> bool

Does the query contain only complements?

If there are no complements, this returns false

pub fn filter_out_complements(&self) -> Self

Filter out all the complements, returning only the positive goals

pub fn join(&self, other: &Query) -> Self

Merge the goals of this query with the goals of another query, returning a new query with all the goals

pub fn remove_provable_complements( &mut self, env: &mut Env<impl Solver>, ) -> bool

Remove all the complements that we can prove are true.

This is done by finding all the complements, and then trying to prove that the complement is false. If we can prove that the complement is false, then we can remove the complement from the query.

If we prove a complement as absurd, then we can return true, indicating that the query is absurd.

pub fn remove_irreducible_negatives(&self, env: &Env<impl Solver>) -> Self

Remove all the complemented goals that cannot be simplified by any of the rules in the environment.

These are inherently negative goals that cannot be proven.

This is a non-mutating version of remove_irreducible_negatives_in_place

pub fn is_empty(&self) -> bool

Is the query empty, i.e. does it have no goals left to prove?

pub fn is_ground_truth(&self) -> bool

Is this query a ground truth query, meaning it is a query that is always true?

A query is ground truth if all the goals are either true or the complement of false. An empty query is also considered ground truth.

pub fn goals(&self) -> impl Iterator<Item = &Term>

Get an iterator over all the goals in the query

pub fn remove_goal(&mut self, goal: &Term)

Remove a goal from the query

pub fn add_positive_goal(&mut self, goal: Term)

Add a positive goal to the query. This means that the goal must be proven true for the query to be true.

pub fn add_negative_goal(&mut self, goal: Term)

Add a negative goal to the query. This means that the goal must be proven false for the query to be true.

pub fn does_contradict_goals(&self, other_goal: &Term) -> bool

Does this query contain a goal that contradicts the given term?

pub fn contains_contradiction(&self) -> bool

Does this query contain contradictory goals?

If false is in the query, then the query is contradictory. If a goal is in the query, and its complement is also in the query, then the query is contradictory.

pub fn used_vars(&self, vars: &mut HashSet<Var>)

Get all the used variables in the query

pub fn has_vars(&self) -> bool

Does the query contain any variables?

pub fn substitute(&mut self, bindings: &HashMap<Var, Term>)

Substitute all occurrences of some variables in the query with the given terms

pub fn reduce<S>(&self, env: &Env<S>) -> Self
where S: Solver,

Simplify the query by substituting an environment’s variable bindings into the query’s goals. This makes the query’s goals more concrete.

This does not mutate the query, but returns a new query with the substituted goals.

pub fn reduce_in_place(&mut self, env: &Env<impl Solver>)

Simplify the query by substituting an environment’s variable bindings into the query’s goals. This makes the query’s goals more concrete.

This mutates the query in place.