reckon
Struct SearchConfig

pub struct SearchConfig<S>
where
    S: Solver,
{

    pub max_search_depth: Option<usize>,
    pub max_search_width: Option<usize>,
    pub step_limit: Option<usize>,
    pub prune: bool,
    pub traversal: Traversal,
    pub require_head_match: bool,
    pub queue_sorter: Option<Arc<dyn Fn(&mut VecDeque<(Env<S>, Query, usize)>)>>,
    pub sort_after_steps: usize,
    pub reduce_query: bool,
    pub solution_limit: usize,
    pub clean_memoization: bool,
    pub stop_after_first_goal: bool,
    pub enable_sorting: bool,

}

A configuration for a search.

This struct contains all the parameters that can be used to configure the search for solutions.

max_search_depth: The maximum depth to search to. If None, the search will continue until a solution is found or the step limit is reached.

max_search_width: The maximum width of the search โ this constrains how many subsearches can be performed for each rule application. If the width is reached, the possible paths remaining for a given query will be pruned.

step_limit: The maximum number of steps to perform in the search. If None, the search will continue until a solution is found or the depth limit is reached.

prune: Whether to prune the search space by removing redundant variables and simplifying queries while searching. This likely decreases the search space, but may increase the time taken to find a solution.

traversal: The traversal strategy to use for the search.

require_head_match: Whether to require that a rule applies to the first goal in a query before applying it to the rest of the goals.

queue_sorter: A function that can be used to sort the queue of queries after a certain number of steps. This can be used to prioritize certain queries.

sort_after_steps: The number of steps to perform before sorting the queue of queries.

reduce_query: Whether to reduce the query after each rule application. This can be used to simplify the query and reduce the search space, but may increase the time taken to find a solution.

solution_limit: The maximum number of solutions to find. If this limit is reached, the search will stop.

clean_memoization: Whether to clean the memoization cache after the full search. After the full search is performed, the final solution is retained and memoized, but the sub-solutions are removed from the cache.

Fields
max_search_depth: Option<usize>

The maximum depth to search to. If None, the search will continue until a solution is found or the step limit is reached.

max_search_width: Option<usize>

The maximum width of the search โ this constrains how many subsearches can be performed for each rule application.

step_limit: Option<usize>

The maximum number of steps to perform in the search. If None, the search will continue until a solution is found or the depth limit is reached.

prune: bool

Whether to prune the search space by removing redundant variables and simplifying queries while searching. This likely decreases the search space,

traversal: Traversal

The traversal strategy to use for the search.

require_head_match: bool

Whether to require that a rule applies to the first goal in a query before applying it to the rest of the goals.

queue_sorter: Option<Arc<dyn Fn(&mut VecDeque<(Env<S>, Query, usize)>)>>

A function that can be used to sort the queue of queries after a certain number of steps. This can be used to prioritize certain queries.

sort_after_steps: usize

The number of steps to perform before sorting the queue of queries.

reduce_query: bool

Whether to reduce the query after each rule application. This can be used to simplify the query and reduce the search space, but may increase the time taken to find a solution.

solution_limit: usize

The maximum number of solutions to find. If this limit is reached, the search will stop.

clean_memoization: bool

Whether to clean the memoization cache after the full search. After the full search is performed,

stop_after_first_goal: bool

Stop application after the first goal in the query

enable_sorting: bool

Enable the use of sorting the queue of queries after a certain number of steps.

Implementations

impl<S> SearchConfig<S>
where S: Solver,

pub fn permissive() -> Self

Create a new search configuration with permissive defaults.

This configuration allows for a wide search space, with no limits on depth, width, or steps.

It is not recommended to use this configuration for most searches, as it may lead to long search times and high memory usage.

If you use this configuration, you should set limits on the search depth, width, and steps.

pub fn parse<'a, 'b>(&'a mut self, s: &'b str) -> Result<&'b str, String>

Parse a search configuration from a string.

This will parse the new options from the string and update the search configuration.

pub fn with_solution_limit(self, solution_limit: usize) -> Self

Create a new search configuration with the given maximum search depth.

pub fn with_traversal(self, traversal: Traversal) -> Self

Create a new search configuration with the given traversal strategy.

pub fn with_sorting_enabled(self, enable_sort: bool) -> Self

Enable sorting the queue of queries after a certain number of steps.

pub fn with_clean_memoization(self, clean_memoization: bool) -> Self

Create a new search configuration with the given clean memoization setting.

pub fn with_stop_after_first_goal(self, stop_after_first_goal: bool) -> Self

Create a new search configuration with the given stop after first goal setting.

pub fn with_reduce_query(self, reduce_query: bool) -> Self

Create a new search configuration with the given reduce query setting.

pub fn with_require_rule_head_match(self, require_head_match: bool) -> Self

Create a new search configuration with the given require head match setting.

pub fn with_depth_limit(self, max_search_depth: usize) -> Self

Create a new search configuration with the given depth limit.

pub fn with_width_limit(self, max_search_width: usize) -> Self

Create a new search configuration with the given width limit.

pub fn with_step_limit(self, step_limit: usize) -> Self

Create a new search configuration with the given step limit.

pub fn with_sorter<K>( self, after_steps: usize, sorter: impl Fn(&Env<S>, &Query) -> K + 'static, ) -> Self
where K: Ord,

Create a new search configuration with the given sorter function, and the number of steps to sort after.

pub fn sort_queue(&self, queue: &mut VecDeque<(Env<S>, Query, usize)>)

Sort the queue of queries using the sorter function.

pub fn with_pruning(self, prune: bool) -> Self

Create a new search configuration with the given pruning setting.

pub fn can_prune(&self) -> bool

Check if the search configuration allows pruning.

pub fn can_search_deeper(&self, depth: usize) -> bool

Check if the search configuration allows deeper searches.

pub fn can_search_wider(&self, width: usize) -> bool

Check if the search configuration allows wider searches.

pub fn can_perform_step(&self, steps: usize) -> bool

Check if the search configuration allows more steps to be performed.