reckon
Trait Solver

pub trait Solver:
    Default
    + Clone
    + Debug
    + Eq {
    // Provided methods
    fn reset(&mut self) { ... }
    fn save_rule_application(
        &mut self,
        _env: Env<Self>,
        _before_query: &Query,
        _new_env: Env<Self>,
        _new_query: &Query,
    ) { ... }
    fn use_saved_rule_application(
        &self,
        _env: &mut Env<Self>,
        _query: &mut Query,
    ) -> bool { ... }
    fn save_solutions(
        &mut self,
        _env: Env<Self>,
        _query: Query,
        _solution: HashSet<Solution>,
    ) { ... }
    fn use_saved_solutions(
        &self,
        _env: &Env<Self>,
        _query: &Query,
    ) -> Option<HashSet<Solution>> { ... }
}

A solver that saves and reuses rule applications and solutions, according to whatever strategy it implements.

Provided Methods

fn reset(&mut self)

Reset the solver’s state, clearing any saved rule applications or solutions

fn save_rule_application( &mut self, _env: Env<Self>, _before_query: &Query, _new_env: Env<Self>, _new_query: &Query, )

Save a rule application for later reuse

fn use_saved_rule_application( &self, _env: &mut Env<Self>, _query: &mut Query, ) -> bool

Attempt to use a saved rule application, returning true if one was found

fn save_solutions( &mut self, _env: Env<Self>, _query: Query, _solution: HashSet<Solution>, )

Save a set of solutions for later reuse

fn use_saved_solutions( &self, _env: &Env<Self>, _query: &Query, ) -> Option<HashSet<Solution>>

Attempt to use saved solutions, returning them if found

Object Safety

This trait is not object safe.